Add test for scenario loading
